projects/alpha_sentiment/pyproject.toml
[project]
name = "alpha-sentiment"
version = "0.1.0"
description = "股票情绪分析终端 - 基于 Tushare 和 DeepSeek API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.104.1",
  "uvicorn[standard]>=0.24.0",
  "gunicorn>=21.2.0",
  "akshare>=1.12.0",
  "pyqlib>=0.9.0",
  "openai>=1.6.0",
  "apscheduler>=3.10.4",
  "pydantic>=2.5.2",
  "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
  "pytest>=7.0.0",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=4.0.0",
  "httpx>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
asyncio_mode = "auto"
markers = [
  "integration: marks tests as integration tests (require network)",
  "e2e: marks tests as end-to-end tests (require running server)",
  "slow: marks tests as slow running",
]
